wall running fully fixed
wall running can now be performed on any wall with the correct camera tilt and wall jump. I also added a check to make sure the player is looking the right direction. If they are looking directly at the wall or too far away from the wall, the wall run will stop